Adding base class for repoManagerFactory
