+++ server/gitrest/packages/gitrest-base/src/test/routes.spec.ts
@@ -218,6 +218,7 @@ testModes.forEach((mode) => {
             return new IsomorphicGitManagerFactory(
                 testUtils.defaultProvider.get("storageDir"),
                 fileSystemManagerFactory,
+                externalStorageManager,
                 testMode.repoPerDocEnabled,
             );
         }
