+++ server/gitrest/packages/gitrest-base/src/utils/isomorphicgitManager.ts
@@ -3,24 +3,22 @@
  * Licensed under the MIT License.
  */
 
-import { E_TIMEOUT, Mutex, MutexInterface, withTimeout } from "async-mutex";
 import * as isomorphicGit from "isomorphic-git";
 import type * as resources from "@fluidframework/gitresources";
 import { NetworkError } from "@fluidframework/server-services-client";
 import { Lumberjack } from "@fluidframework/server-services-telemetry";
+import { IExternalStorageManager } from "../externalStorageManager";
 import * as helpers from "./helpers";
 import * as conversions from "./isomorphicgitConversions";
 import {
-    IRepositoryManagerFactory,
     IExternalWriterConfig,
     IRepositoryManager,
     IFileSystemManager,
     IFileSystemManagerFactory,
-    IRepoManagerParams,
     IStorageDirectoryConfig,
     BaseGitRestTelemetryProperties,
-    Constants,
 } from "./definitions";
+import { RepositoryManagerFactoryBase } from "./repositoryManagerFactoryBase";
 
 export class IsomorphicGitRepositoryManager implements IRepositoryManager {
     constructor(
@@ -359,167 +357,41 @@ export class IsomorphicGitRepositoryManager implements IRepositoryManager {
     }
 }
 
-export class IsomorphicGitManagerFactory implements IRepositoryManagerFactory {
-    private readonly repositoryCache = new Set<string>();
-    private readonly mutexes = new Map<string, MutexInterface>();
-    private readonly internalHandler: (
-        params: IRepoManagerParams,
-        onRepoNotExists: (args?: any) => Promise<void> | never,
-        shouldUseMutex: boolean) => Promise<IsomorphicGitRepositoryManager>;
-
+export class IsomorphicGitManagerFactory extends RepositoryManagerFactoryBase<void> {
     constructor(
-        private readonly storageDirectoryConfig: IStorageDirectoryConfig,
-        private readonly fileSystemManagerFactory: IFileSystemManagerFactory,
+        storageDirectoryConfig: IStorageDirectoryConfig,
+        fileSystemManagerFactory: IFileSystemManagerFactory,
+        externalStorageManager: IExternalStorageManager,
         repoPerDocEnabled: boolean,
     ) {
-        if (repoPerDocEnabled) {
-            this.internalHandler = this.repoPerDocInternalHandler.bind(this);
-        } else {
-            this.internalHandler = this.repoPerTenantInternalHandler.bind(this);
-        }
+        super(storageDirectoryConfig, fileSystemManagerFactory, externalStorageManager, repoPerDocEnabled);
     }
 
-    public async create(params: IRepoManagerParams): Promise<IsomorphicGitRepositoryManager> {
-        const onRepoNotExists = async (args?: any) => {
-            await isomorphicGit.init({
-                fs: args?.fileSystemManager,
-                gitdir: args?.directoryPath,
-                bare: true,
-            });
-            this.repositoryCache.add(args?.repoPath);
-            Lumberjack.info(
-                "Created a new repo",
-                {
-                    ...(args?.lumberjackBaseProperties),
-                    [BaseGitRestTelemetryProperties.directoryPath]: args?.directoryPath,
-                });
-        };
-
-        return this.internalHandler(params, onRepoNotExists, true);
+    protected async initGitRepo(fs: IFileSystemManager, gitdir: string): Promise<void> {
+        return isomorphicGit.init({
+            fs,
+            gitdir,
+            bare: true,
+        });
     }
 
-    public async open(params: IRepoManagerParams): Promise<IsomorphicGitRepositoryManager> {
-        const onRepoNotExists = (args?: any) => {
-            Lumberjack.error(
-                `Repo does not exist ${args?.directoryPath}`,
-                {
-                    ...(args?.lumberjackBaseProperties),
-                    [BaseGitRestTelemetryProperties.directoryPath]: args?.directoryPath,
-                });
-                // services-client/getOrCreateRepository depends on a 400 response code
-                throw new NetworkError(400, `Repo does not exist ${args?.directoryPath}`);
-        };
-
-        return this.internalHandler(params, onRepoNotExists, false);
+    protected async openGitRepo(gitdir: string): Promise<void> {
+        return;
     }
 
-    public async openOrCreate(params: IRepoManagerParams): Promise<IsomorphicGitRepositoryManager> {
-        try {
-            return await this.open(params);
-        } catch (error: any) {
-            if (error instanceof Error &&
-                error?.name === "NetworkError" &&
-                (error as NetworkError)?.code === 400) {
-                    return this.create(params);
-            }
-            throw error;
-        }
-    }
-
-    private async repoPerDocInternalHandler(
-        params: IRepoManagerParams,
-        onRepoNotExists: (args?: any) => Promise<void> | never,
-        shouldUseMutex: boolean): Promise<IsomorphicGitRepositoryManager> {
-        if (!params.storageRoutingId?.tenantId || !params.storageRoutingId?.documentId) {
-            throw new NetworkError(400, `Invalid ${Constants.StorageRoutingIdHeader} header`);
-        }
-
-        const repoPath = helpers.getRepoPath(
-            params.storageRoutingId.tenantId,
-            params.storageRoutingId.documentId,
-            this.storageDirectoryConfig.useRepoOwner ? params.repoOwner : undefined);
-        const directoryPath = helpers.getGitDirectory(repoPath, this.storageDirectoryConfig.baseDir);
-        const repoName = `${params.storageRoutingId.tenantId}/${params.storageRoutingId.documentId}`;
-
-        return this.internalHandlerCore(
-            params,
-            repoPath,
-            directoryPath,
-            repoName,
-            onRepoNotExists,
-            shouldUseMutex);
-    }
-
-    private async repoPerTenantInternalHandler(
-        params: IRepoManagerParams,
-        onRepoNotExists: (args?: any) => Promise<void> | never,
-        shouldUseMutex: boolean): Promise<IsomorphicGitRepositoryManager> {
-        const repoPath = helpers.getRepoPath(
-            params.repoName,
-            undefined,
-            this.storageDirectoryConfig.useRepoOwner ? params.repoOwner : undefined);
-        const directoryPath = helpers.getGitDirectory(repoPath, this.storageDirectoryConfig.baseDir);
-
-        return this.internalHandlerCore(
-            params,
-            repoPath,
-            directoryPath,
-            params.repoName,
-            onRepoNotExists,
-            shouldUseMutex);
-    }
-
-    private async internalHandlerCore(
-        params: IRepoManagerParams,
-        repoPath: string,
-        directoryPath: string,
+    protected createRepoManager(
+        fileSystemManager: IFileSystemManager,
+        repoOwner: string,
         repoName: string,
-        onRepoNotExists: (args?: any) => Promise<void> | never,
-        shouldUseMutex: boolean): Promise<IsomorphicGitRepositoryManager> {
-        const lumberjackBaseProperties = helpers.getLumberjackBasePropertiesFromRepoManagerParams(params);
-        const fileSystemManager = this.fileSystemManagerFactory.create(params.fileSystemManagerParams);
-
-        // We define the function below to be able to call it either on its own or within the mutex.
-        const action = async () => {
-            if (!(this.repositoryCache.has(repoPath))) {
-                const repoExists = await helpers.exists(fileSystemManager, directoryPath);
-                if (!repoExists || !repoExists.isDirectory()) {
-                    await onRepoNotExists({
-                        fileSystemManager,
-                        repoPath,
-                        directoryPath,
-                        lumberjackBaseProperties });
-                } else {
-                    this.repositoryCache.add(repoPath);
-                }
-            }
-
-            const repoManager = new IsomorphicGitRepositoryManager(
+        repo: void,
+        gitdir: string,
+        externalStorageManager: IExternalStorageManager,
+        lumberjackBaseProperties: Record<string, any>): IRepositoryManager {
+            return new IsomorphicGitRepositoryManager(
                 fileSystemManager,
-                params.repoOwner,
+                repoOwner,
                 repoName,
-                directoryPath,
+                gitdir,
                 lumberjackBaseProperties);
-            return repoManager;
-        };
-
-        if (shouldUseMutex) {
-            if (!this.mutexes.has(repoName)) {
-                this.mutexes.set(repoName, withTimeout(new Mutex(), 10000));
-            }
-            const mutex = this.mutexes.get(repoName);
-            try {
-                return mutex.runExclusive(async () => {
-                    return action();
-                });
-            } catch (e: any) {
-                if (e === E_TIMEOUT) {
-                    throw new NetworkError(500, "Could not complete action due to mutex timeout.");
-                }
-                throw new NetworkError(500, `Unknown error when trying to run action:  ${e?.message}`);
-            }
-        } else {
-            return action();
-        }
     }
 }
